lmb-tables: restructure utils and using meeting start date instead of current date
