+++ app/components/inauguration-meeting/synchronization.js
@@ -11,9 +11,9 @@ import ENV from 'frontend-gelinkt-notuleren/config/environment';
 import { headlessProsemirror } from '../../utils/meeting/headless-prosemirror';
 import InaugurationMeetingSynchronizationToast from './synchronization-toast';
 import { syncDocument } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/mandatee-table-plugin';
-import { MANDATEE_TABLE_SAMPLE_CONFIG } from '../../config/mandatee-table-config';
 import SaySerializer from '@lblod/ember-rdfa-editor/core/say-serializer';
 import { getOwner } from '@ember/application';
+import { mandateeTableConfigIVGR } from '../../config/mandatee-table-config';
 
 export default class InaugurationMeetingSynchronizationComponent extends Component {
   @service toaster;
@@ -196,7 +196,7 @@ export default class InaugurationMeetingSynchronizationComponent extends Compone
     const initialState = headlessProsemirror(html, getOwner(this));
     const syncedState = await syncDocument(
       initialState,
-      MANDATEE_TABLE_SAMPLE_CONFIG,
+      mandateeTableConfigIVGR(this.meeting),
     );
     const serializer = SaySerializer.fromSchema(
       syncedState.schema,
