+++ app/utils/sparql.js
@@ -1,0 +1,11 @@
+/**
+ *
+ * Converts a sparql binding to a simple object contain a mapping of the binding keys to their values
+ */
+export function bindingToObject(binding) {
+  const result = {};
+  for (let key in binding) {
+    result[key] = binding[key].value;
+  }
+  return result;
+}
